Add unit tests to AirportQueryBuilders.
Added unit tests for SourceAirportQueryBuilder.
Fixed a bug where IsModelNotRun does not account for when the only
locations of the event are country only. Added the condition to the
query. Note that in production, no airports will be in the table
anyways if the model was not run, so it will always be an empty set
returned, but this is to future-proof in case we do have airports
even if we don't run the model.

Added unit tests for DestinationAirportQueryBuilder.
Added the Moq framework to allow overriding of Stored Procedure raw
sql interpolated calls. Created the extension for DbSet to provide
a method that injects a custom action while receiving the original
SQL query call as arguments. In DestinationAirportQueryBuilderTest,
we use a switch on the provided geoname to return pre-defined set of
data based on the geoname id.

Fixed a bug to IsModelNotRun similar to SourceAirportQueryBuilder.
Also fixed the longitude and latitude to read from Stations.
Also added comments about the complex dataset.

Merged in feature/unit-test-data (pull request #679)
Approved-by: Mikhail Temkine <mikhail.temkine@bluedot.global>