Fix application crash when saving user location.
The SP being called to lookup the geoname looks in ActiveGeonames,
which does not have the newly searched Geoname. Updated the SP to not
use the ActiveGeonames table (new Products DB will not have this) so
that there will be less potential for failure when calling this SP if
the ActiveGeonames does not have the requested geoname id.

Fixes PT-1293.
Merged in bugfix/PT-1293-update-user-location-crash (pull request #689)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>