Update APIs to support new Export Airport changes.
The Event Details for exportation has been updated so that the export
airports are similar to the import (instead of volume, show risk). To
allow the front-end to show this, the API now returns the exportation
risk and ordering by risk, instead of volume.

Part of PT-1307.
Merged in feature/PT-1307-export-airport-sorting (pull request #691)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>
Approved-by: Mikhail Temkine <mikhail.temkine@bluedot.global>