Update notifications to use new risk categories.
In the past, we show the min to max probabilities in the New Event
and Weekly email as an interval. This has now been categorized into
buckets (i.e. Low, Moderate, High...etc) with sub-text showing what
the range for that category consists of.

Updated the getInterval helper function and related unit tests.
Also updated the sorting of the events to be using the average.

Implements PT-1302.
Merged in feature/PT-1302-trvlcats-notification (pull request #699)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>