Prevent caching of index.html in react app.
The index.html was being cached by the browser. This file has the
references to which resources (JS and CSS) version to use, and if it
is old, uses the old JS and CSS files.

Fixed by setting a no cache outbound rule to the index.html file so
that browsers will not cache it and always load a new one for latest.

Note: Existing index.html are not expired may still cause issues
until the user loads this new index.html from the server.

Fixes PT-1295.
Merged in bugfix/PT-1295-index-html-cache-issue (pull request #726)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>