Added IATA and Innovata year data to configs.
Previously, IATA was read from the latest year in the GridStation
table due to the fact that the year is the same for all months. This
is no longer true as the latest month is the current year. The UI now
shows 2 different years, 1 for IATA and 1 for Innovata. Both of these
properties have been added into the DB with post-deployment script,
and returned via the API.

Part of PT-1366.
Merged in feature/PT-1366-transpar-source-update (pull request #735)
Approved-by: Mikhail Temkine <mikhail.temkine@bluedot.global>