Update Insights code to use UserProfile table.
Updated the Entity Framework to include the latest schema.

The User Type is available via the UserProfile table which is more
robust to use than User Role (not enforced to exactly 1). Updated the
codebase of insights to join AspNetUsers to UserProfile so that both
table properties are available when looking up the user. Any action
related to relevance settings will use the UserProfile to update.

Note that there are triggers that update UserProfile with AspNetUsers
so user properties will still be updated to that table until full
migration.

Also updated unit tests to be related to User Type instead of role.

Part of PT-1250 and PT-1371.
Merged in feature/PT-1250-update-insights-code (pull request #765)
Approved-by: Mikhail Temkine <mikhail.temkine@bluedot.global>