Add skeleton for implementing PROXPAR
Calculation of proximal case counts are currently implemented in `usp_ZebraDiseaseGetLocalCaseCount`. This needs to be modified for PROXPAR since the SP does not differentiate the case counts between what's in the user's location vs what is near the user's location. The logic will be moved from SP to C# and this commit provides the skeleton to do this task.

Part of PT-1298
Merged in feature/PT-1298-proximal-location-structure (pull request #774)
Approved-by: Kevin Chu