Added function for proximal event locations.
The new function will provide a breakdown for any events that are in
or proximal to the input geoname id. All event locations of those
events are returned with a flag IsWithinLocation so that any nesting
calculation can still be done on the full set of event locations,
and can subsequently be filtered out by the flag.

Part of PT-1298.
Merged in feature/PT-1298-proximal-event-location-func (pull request #784)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>