Added filtering of empty strings for AOI ids.
Although this should never happen, a few user's AOI was an empty
string due to a bug in saving Custom Settings (already addressed).
But due to this empty string, the emails were failing as when split
by ',' the result is an empty string that could not be casted to an
integer, hence causing total failure of the email sending service.

Added a check to remove any invalid geoname ids, to prevent any
future failures due to no AOI.

Fixes PT-1431 and PT-1432.
Merged in bugfix/PT-1431-empty-aoi-email-crash (pull request #792)
Approved-by: Mikhail Temkine <mikhail.temkine@bluedot.global>