Group proximal locations by geoname ID
This should NOT happen in production since there should be no disease reported in a location that spans multiple events but done in side of caution. This commit sums cases when `ufn_GetProximalEventLocations` returns more than 1 row for a location.

Part of PT-1298
Merged in feature/PT-1298-multiple-location-edge-case (pull request #808)
Approved-by: Kevin Chu