Remove extra calls to update event location history.
Previously, the event location history table of type 1 (proximal) was
updated separately from the Proximal Email call. It was in both the
ZebraEventUpdateController as well as Surveillance external API call.

Since the change has been made to the Proximal Notification Service
so the service makes the call to update the history after sending
the proximal email, the extra async calls are now obsolete and is
causing extra updates to the history at unexpected times.

Fixes PT-1454.
Merged in hotfix/PT-1454-proximal-email-not-fired (pull request #814)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>