Remove failing unit tests due to new transaction.
The Zebra Event Update API has now a transaction clause added around
the core logic. Since the unit test Mock Context does not provide the
transaction ability, the unit tests are all failing.

There is no easy way to address this currently, so will remove the
set of unit tests that are testing this piece of code. There are also
future plans to migrate this logic off of Zebra API.