Added SP that calls the table-valued func.
The function is having performance issues when called from EF and is
hanging. Calling the function from SSMS is fine and is fast.

Created a SP that essentially acts as a wrapper around the function
so that the EF code can call the SP (no issues).

Merged in bugfix/function-perf-issue (pull request #819)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>