Add functions for nested event locations.
Since there is a bug associated to the EventNestedLocation table
where the data is not properly nested in some cases, and the proper
fix is to have full historical incremental deltas between data points
as a pre-cursor to the Disease Activity Stop gap implementation, two
new functions that compute the nested event locations at a given time
is created.

The new function will replace any usage of the bugged view, and is
written in a way where nesting is computed for the event locations
at a specific date when executed.

Fixes for PT-1739.
Merged in bugfix/nested-event-location-bugged (pull request #832)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>