Remove waiting of Zebra Event Update in Surveillance.
Previously, when pushing an event to Zebra, Surveillance will wait
for a response before continuing due to the fact that the call to
send Proximal/New Event emails are made by Surveillance. Moved this
call to Zebra API so that after risk is calculated, the call to the
Notification API is made to fire the emails. On Surveillance end, the
call to Zebra Event Update API is now fired without waiting.

Also added a few more logging messages to help diagnose where the
slow parts of risk calculation are.

Fixes PT-1477.
Merged in bugfix/PT-1477-surveillance-ignore-response (pull request #849)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>