Move Proximal Case response to separate API.
For places that want to display the proximal case count distribution,
the data will not be returned with the Get Events/Event API due to
the calls slowing down loading of the properties. The UI will show
loading before providing the proximal data.

Fixes PT-1476.
Merged in bugfix/PT-1476-proxpar-separate-api (pull request #851)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>