Fix large table join causing slow query.
The EventSourceGridSpreadMd contains a large amount of rows and left
joining (when using .Include) causes the Event row to be duplicated
in SQL, which is very slow when the data in the Event table is large.

Fixed by moving the loading of EventSourceGridSpreadMd to be another
call instead of using the join.

Merged in hotfix/event-details-performance (pull request #859)
Approved-by: Vivian Hu <vivian.hu@bluedot.global>