Fix ZebraEventUpdateController bugs
Defaulted to 0 cases since Poisson distribution cannot take lambda
as 0.
Added flag to allow forceRefresh flag to bypass isCalculating flag
so that if the event is stuck in isCalculating state, the force will
allow re-calculation to attempt to fix it.
Also added a secondary reset of the flag in case the SP does not
properly reset the calculating flag.